style: 파일 위치 변경 및 분리
