+++ lib/240325_homework/model/todo.dart
@@ -1,0 +1,50 @@
+class Todo {
+  final int userId;
+  final int id;
+  final String title;
+  final bool completed;
+
+  Todo(this.userId, this.id, this.title, this.completed);
+
+  Todo.fromJson(Map<String, dynamic> json)
+      : userId = json['userId'],
+        id = json['id'],
+        title = json['title'],
+        completed = json['completed'];
+
+  Map<String, dynamic> toJson() =>
+      {'userId': userId, 'id': id, 'title': title, 'completed': completed};
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is Todo &&
+          runtimeType == other.runtimeType &&
+          userId == other.userId &&
+          id == other.id &&
+          title == other.title &&
+          completed == other.completed;
+
+  @override
+  int get hashCode =>
+      userId.hashCode ^ id.hashCode ^ title.hashCode ^ completed.hashCode;
+
+  @override
+  String toString() {
+    return 'Todo{userId: $userId, id: $id, title: $title, completed: $completed}';
+  }
+
+  Todo copyWith({
+    int? userId,
+    int? id,
+    String? title,
+    bool? completed,
+  }) {
+    return Todo(
+      userId ?? this.userId,
+      id ?? this.id,
+      title ?? this.title,
+      completed ?? this.completed,
+    );
+  }
+}
