+++ lib/240325_homework/user_api.dart
@@ -2,6 +2,8 @@ import 'dart:convert';
 
 import 'package:http/http.dart' as http;
 
+import './model/user.dart';
+
 void main() async {
   UserApi user = UserApi();
   await user.getUsers();
@@ -33,241 +35,5 @@ class UserApi {
     } else {
       throw Exception('Response 에러');
     }
-  }
-}
-
-class User {
-  final int id;
-  final String name;
-  final String username;
-  final String email;
-  final Address address;
-  final String phone;
-  final String website;
-  final Company company;
-
-  User(this.id, this.name, this.username, this.email, this.address, this.phone,
-      this.website, this.company);
-
-  User.fromJson(Map<String, dynamic> json)
-      : id = json['id'],
-        name = json['name'],
-        username = json['username'],
-        email = json['email'],
-        address = Address.fromJson(json['address']),
-        phone = json['phone'],
-        website = json['website'],
-        company = Company.fromJson(json['company']);
-
-  Map<String, dynamic> toJson() => {
-        'id': id,
-        'name': name,
-        'username': username,
-        'email': email,
-        'address': address.toJson(),
-        'phone': phone,
-        'website': website,
-        'company': company.toJson(),
-      };
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is User &&
-          runtimeType == other.runtimeType &&
-          id == other.id &&
-          name == other.name &&
-          username == other.username &&
-          email == other.email &&
-          address == other.address &&
-          phone == other.phone &&
-          website == other.website &&
-          company == other.company;
-
-  @override
-  int get hashCode =>
-      id.hashCode ^
-      name.hashCode ^
-      username.hashCode ^
-      email.hashCode ^
-      address.hashCode ^
-      phone.hashCode ^
-      website.hashCode ^
-      company.hashCode;
-
-  @override
-  String toString() {
-    return 'User{id: $id, name: $name, username: $username, email: $email, address: $address, phone: $phone, website: $website, company: $company}';
-  }
-
-  User copyWith({
-    int? id,
-    String? name,
-    String? username,
-    String? email,
-    Address? address,
-    String? phone,
-    String? website,
-    Company? company,
-  }) {
-    return User(
-      id ?? this.id,
-      name ?? this.name,
-      username ?? this.username,
-      email ?? this.email,
-      address ?? this.address,
-      phone ?? this.phone,
-      website ?? this.website,
-      company ?? this.company,
-    );
-  }
-}
-
-class Address {
-  final String street;
-  final String suite;
-  final String city;
-  final String zipcode;
-  final Geo geo;
-
-  Address(this.street, this.suite, this.city, this.zipcode, this.geo);
-
-  Address.fromJson(Map<String, dynamic> json)
-      : street = json['street'],
-        suite = json['suite'],
-        city = json['city'],
-        zipcode = json['zipcode'],
-        geo = Geo.fromJson(json['geo']);
-
-  Map<String, dynamic> toJson() => {
-        'street': street,
-        'suite': suite,
-        'city': city,
-        'zipcode': zipcode,
-        'geo': geo.toJson(),
-      };
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is Address &&
-          runtimeType == other.runtimeType &&
-          street == other.street &&
-          suite == other.suite &&
-          city == other.city &&
-          zipcode == other.zipcode &&
-          geo == other.geo;
-
-  @override
-  int get hashCode =>
-      street.hashCode ^
-      suite.hashCode ^
-      city.hashCode ^
-      zipcode.hashCode ^
-      geo.hashCode;
-
-  @override
-  String toString() {
-    return 'Address{street: $street, suite: $suite, city: $city, zipcode: $zipcode, geo: $geo}';
-  }
-
-  Address copyWith({
-    String? street,
-    String? suite,
-    String? city,
-    String? zipcode,
-    Geo? geo,
-  }) {
-    return Address(
-      street ?? this.street,
-      suite ?? this.suite,
-      city ?? this.city,
-      zipcode ?? this.zipcode,
-      geo ?? this.geo,
-    );
-  }
-}
-
-class Geo {
-  final String lat;
-  final String lng;
-
-  Geo(this.lat, this.lng);
-
-  Geo.fromJson(Map<String, dynamic> json)
-      : lat = json['lat'],
-        lng = json['lng'];
-
-  Map<String, dynamic> toJson() => {
-        'lat': lat,
-        'lng': lng,
-      };
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is Geo &&
-          runtimeType == other.runtimeType &&
-          lat == other.lat &&
-          lng == other.lng;
-
-  @override
-  int get hashCode => lat.hashCode ^ lng.hashCode;
-
-  @override
-  String toString() {
-    return 'Geo{lat: $lat, lng: $lng}';
-  }
-
-  Geo copyWith({
-    String? lat,
-    String? lng,
-  }) {
-    return Geo(
-      lat ?? this.lat,
-      lng ?? this.lng,
-    );
-  }
-}
-
-class Company {
-  final String name;
-  final String catchPhrase;
-  final String bs;
-
-  Company(this.name, this.catchPhrase, this.bs);
-
-  Company.fromJson(Map<String, dynamic> json)
-      : name = json['name'],
-        catchPhrase = json['catchPhrase'],
-        bs = json['bs'];
-
-  Map<String, dynamic> toJson() =>
-      {'name': name, 'catchPhrase': catchPhrase, 'bs': bs};
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is Company &&
-          runtimeType == other.runtimeType &&
-          name == other.name &&
-          catchPhrase == other.catchPhrase &&
-          bs == other.bs;
-
-  @override
-  int get hashCode => name.hashCode ^ catchPhrase.hashCode ^ bs.hashCode;
-
-  @override
-  String toString() {
-    return 'Company{name: $name, catchPhrase: $catchPhrase, bs: $bs}';
-  }
-
-  Company copyWith({
-    String? name,
-    String? catchPhrase,
-    String? bs,
-  }) {
-    return Company(
-        name ?? this.name, catchPhrase ?? this.catchPhrase, bs ?? this.bs);
   }
 }
